requirements.txt
wrapt==1.10.11
wheel==0.31.0
Werkzeug==0.14.1
urllib3==1.22
update-checker==0.16
tornado==5.0.1
tflearn==0.3.2
termcolor==1.1.0
tensorflow-gpu==1.7.0
tensorboard==1.7.0
six==1.11.0
setuptools==39.0.1
scipy==1.0.0
scikit-learn==0.19.1
requests==2.18.4
PyYAML==3.12
pytz==2018.3
python-dateutil==2.7.0
pyparsing==2.2.0
pylint==1.8.3
pydot==1.2.4
pycodestyle==2.3.1
protobuf==3.5.2.post1
prawcore==0.14.0
praw==5.4.0
pip==9.0.3
Pillow==5.1.0
pep8==1.7.1
pandas==0.22.0
packaging==17.1
numpy==1.14.2
mccabe==0.6.1
MarkupSafe==1.0
Markdown==2.6.11
lazy-object-proxy==1.3.1
Keras==2.1.5
Jinja2==2.10
isort==4.3.4
idna==2.6
html5lib==0.9999999
h5py==2.7.1
grpcio==1.10.0
graphviz==0.8.2
gast==0.2.0
cycler==0.10.0
chardet==3.0.4
certifi==2018.1.18
bokeh==0.12.15
bleach==1.5.0
autopep8==1.3.4
astroid==1.6.2
astor==0.6.2
absl-py==0.1.13
